+++ starsky/starsky.foundation.import/starsky.foundation.import.csproj
@@ -8,12 +8,12 @@
         <Nullable>enable</Nullable>
         <RootNamespace>starsky.foundation.import</RootNamespace>
         <ImplicitUsings>disable</ImplicitUsings>
-        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     </PropertyGroup>
 
     <ItemGroup>
         <ProjectReference Include="..\starsky.foundation.database\starsky.foundation.database.csproj"/>
         <ProjectReference Include="..\starsky.foundation.geo\starsky.foundation.geo.csproj"/>
+        <ProjectReference Include="..\starsky.foundation.native\starsky.foundation.native.csproj"/>
         <ProjectReference Include="..\starsky.foundation.thumbnailmeta\starsky.foundation.thumbnailmeta.csproj"/>
         <ProjectReference Include="..\starsky.foundation.worker\starsky.foundation.worker.csproj"/>
         <ProjectReference Include="..\starsky.foundation.writemeta\starsky.foundation.writemeta.csproj"/>
